P1 C3: Move capabilities.rb → inventory/capabilities.rb (move-only)
Module path: Legion::LLM::Capabilities → Legion::LLM::Inventory::Capabilities.
Old path is a require_relative forwarder. compat.rb const_missing alias added.
router.rb and router/availability.rb updated to require new path.
Capabilities calls in availability.rb use fully-qualified Legion::LLM::Inventory::Capabilities.
