+++ lib/legion/llm/compat.rb
@@ -28,6 +28,10 @@ module Legion
         # Moved to Legion::LLM::Inventory::Discovery in v0.14.0
         CompatWarning.warn_once('Legion::LLM::Discovery', 'Legion::LLM::Inventory::Discovery')
         Inventory::Discovery
+      when :Capabilities
+        # Moved to Legion::LLM::Inventory::Capabilities in v0.14.0
+        CompatWarning.warn_once('Legion::LLM::Capabilities', 'Legion::LLM::Inventory::Capabilities')
+        Inventory::Capabilities
       when :Pipeline
         CompatWarning.warn_once('Legion::LLM::Pipeline', 'Legion::LLM::Inference')
         Inference
