+++ lib/legion/llm/router.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative 'capabilities'
+require 'legion/llm/inventory/capabilities'
 require_relative 'router/resolution'
 require_relative 'router/rule'
 require_relative 'router/health_tracker'
